Added implementations in BASIC, C++, and Python
Initial implementations are of claim 26 of the ‘375 Patent, which is
the system claim Rader used.  The two Python claims differ in that v1
uses a dictionary for the accounts and v2 uses an Account class.

Points to revisit after claim construction:
- None of the implementations loop over the creation of multiple
exchange obligations, though they could.
- None of the implementations actually include an “exchange
institution” because, with only two parties interacting, it really
cluttered things up unnecessarily.  We should talk about this.
- As of now only the BASIC implementation outputs text; the other two
have it commented.  I thought we could figure out whether the output is
confusing or helpful (like commenting) later.