Create draft I.C.txt
Anna's draft of section I.C.

Note that this was created in Word (I don't have my laptop today).  It contains ONLY I.C, not the rest of the brief, and has never been compiled.  Checking in so that Charles can see it, and possibly add it to his correct tex document.